Removed both baseheader and basepanel classes
