+++ elements/rh-accordion/rh-accordion-header.css
@@ -14,9 +14,17 @@
 }
 
 #heading {
+  font-size: 100%;
+  padding: 0;
+  margin: 0;
   color: var(--rh-color-text-primary-on-light, #151515);
   background-color: var(--_rhds-background-color, var(--rh-color-surface-lightest, #ffffff));
   font-weight: var(--rh-font-weight-heading-medium, 500);
+}
+
+button,
+a {
+  cursor: pointer;
 }
 
 .dark {
@@ -90,9 +98,10 @@
 }
 
 span {
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
   max-width: calc(100% - var(--rh-space-xl, 24px));
-  text-overflow: unset;
-  white-space: normal;
   text-align: start;
 }
 
@@ -112,7 +121,25 @@ span {
   color: var(--_active-text-color);
 }
 
+.toggle,
+.toggle:before,
 .toggle:after {
+  padding: 0;
+  margin: 0;
+}
+
+.toggle {
+  position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: 100%;
+  border: 0;
+}
+
+.toggle:after {
+  content: "";
+  position: absolute;
   inset-block: 0;
   inset-inline-start: 0;
 }
