+++ elements/rh-accordion/rh-accordion-panel.css
@@ -1,4 +1,7 @@
 :host {
+  display: none;
+  overflow: hidden;
+  will-change: height;
   color: var(--rh-color-surface-darkest, #151515);
 
   --_background-color: var(--rh-color-surface-lightest, #ffffff);
@@ -37,6 +40,8 @@
 }
 
 .body {
+  position: relative;
+  overflow: hidden;
   padding:
     var(--_panel-body-padding-block-start, var(--rh-space-md, 8px))
     var(--_panel-body-padding-inline-end, var(--rh-space-lg, 16px))
@@ -49,10 +54,13 @@
 }
 
 :host([fixed]) {
+  overflow-y: auto;
   max-height: 9.375rem;
 }
 
 :host([expanded]) {
+  display: block;
+  position: relative;
   position: unset;
 }
 
@@ -77,6 +85,11 @@
 }
 
 .body:after {
+  content: "";
+  position: absolute;
+  top: 0;
+  bottom: 0;
+  left: 0;
   width: var(--_panel-body--before--Width, var(--rh-border-width-lg, 3px));
   background-color: var(--_panel-body--before-background-color, transparent);
   inset-block: 0;
